(PC-17012)[API] refactor: move public api to their folder
